Create visual structure and buttons
Added two basic divs to the html file, one for the display and one for the buttons. The display div remains blank. The buttons div has been populated with buttons for each of the 10 digits, the 4 operators, the decimal point, and the equals sign, but the buttons have not been placed correctly yet.