Perform two-operand arithmetic operations in console
By opening the developer console, the user can click the numeric, operator, and equals buttons to perform one round of two-operand arithmetic. The system does not reset when the operation is complete, so pressing additional numbers will continue appending new digits to the second operand and perform the same operation with that new second operand if the equals sign is pressed.