Perform multiple arithmetic operations in a row in the console
The program can now perform multiple arithmetic operations in a row inside the developer console, using the result of the previous operation as the first operand in a new operation. At the end of each operation, the firstOperand variable is given the current value of the result variable and converted back to a string, the secondOperand variable is converted back into an empty string, and the opComplete variable is updated to true, reflecting that an operation has already been performed and the first operand will necessarily be the result of that operation.