Add decimal functionality
The user can now input numbers with decimals. If the user tries to input a decimal and the current operand already has a decimal, the console will log that the user is trying to inappropriately add a decimal, and no changes will be made to the operand with that button press.

The display has also been made more visually appealing, with colors added to the backgrounds and buttons.